+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * renamed `current_severity_level` to `severity_level` in `MemoryPressureManager`
 * deprecated `DictionaryView`
+* deprecated `parallel_for`